ZUMO-2355 Rejoin interfaces and implementation. Change gasLimit type from Decimal to int
