+++ index.ts
@@ -1,5 +1,11 @@
-import zumokit from './src/models/ZumoKit';
+import zumokit from './src/ZumoKit';
 
 export default zumokit;
-export * from './src/types/exported';
-export * from './src/interfaces';
+
+export * from 'zumokit/src/interfaces/public';
+export * from 'zumokit/src/models';
+export { User } from './src/User';
+export { Utils } from './src/Utils';
+export { Wallet } from './src/Wallet';
+export { ZumoKit } from './src/ZumoKit';
+export { ZumoKitError } from './src/ZumoKitError';
