+++ src/User.ts
@@ -1,8 +1,5 @@
 import { NativeModules, NativeEventEmitter } from 'react-native';
-import { Account } from 'zumokit/src/models/Account';
-import { AccountFiatProperties } from 'zumokit/src/models/AccountFiatProperties';
-import { AccountDataSnapshot } from 'zumokit/src/models/AccountDataSnapshot';
-import { User as IUser } from '../interfaces';
+import { Account, AccountFiatProperties, AccountDataSnapshot } from 'zumokit/src/models';
 import {
   AccountJSON,
   CurrencyCode,
@@ -10,15 +7,34 @@ import {
   AccountType,
   AccountDataSnapshotJSON,
   FiatCustomerData,
-  UserJSON,
-} from '../types';
+} from 'zumokit/src/interfaces';
 import { Wallet } from './Wallet';
-import { tryCatchProxy } from '../utility/errorProxy';
+import { tryCatchProxy } from './utility/errorProxy';
 
-const { RNZumoKit } = NativeModules;
+const {
+  /** @internal */
+  RNZumoKit,
+} = NativeModules;
 
+/** @internal */
+interface UserJSON {
+  id: string;
+  hasWallet: boolean;
+  accounts: Array<AccountJSON>;
+}
+
+/**
+ * User instance provides methods for managing user wallet and accounts.
+ * <p>
+ * User instance can be obtained via {@link ZumoKit.signIn} method.
+ * <p>
+ * See <a href="https://developers.zumo.money/docs/guides/manage-user-wallet">Manage User Wallet</a>,
+ * <a href="https://developers.zumo.money/docs/guides/create-fiat-account">Create Fiat Account</a> and
+ * <a href="https://developers.zumo.money/docs/guides/view-user-data">View User Data</a>
+ * guides for usage details.
+ */
 @tryCatchProxy
-export class User implements IUser {
+export class User {
   // Listening to account data changes
   private listeningToChanges = false;
 
@@ -31,12 +47,16 @@ export class User implements IUser {
   // Listeners for account data changes
   private accountDataListeners: Array<(snapshots: Array<AccountDataSnapshot>) => void> = [];
 
+  /** User dentifier. */
   id: string;
 
+  /** Indicator if user has wallet. */
   hasWallet: boolean;
 
+  /** User accounts. */
   accounts: Array<Account>;
 
+  /** @internal */
   constructor(json: UserJSON) {
     this.id = json.id;
     this.hasWallet = json.hasWallet;
@@ -52,30 +72,62 @@ export class User implements IUser {
     });
   }
 
+  /**
+   * Create user wallet seeded by provided mnemonic and encrypted with user's password.
+   * <p>
+   * Mnemonic can be generated by {@link Utils.generateMnemonic} utility method.
+   * @param  mnemonic       mnemonic seed phrase
+   * @param  password       user provided password
+   */
   async createWallet(mnemonic: string, password: string) {
     await RNZumoKit.createWallet(mnemonic, password);
     this.hasWallet = true;
     return new Wallet();
   }
 
+  /**
+   * Recover user wallet with mnemonic seed phrase corresponding to user's wallet.
+   * This can be used if user forgets his password or wants to change his wallet password.
+   * @param  mnemonic       mnemonic seed phrase corresponding to user's wallet
+   * @param  password       user provided password
+   */
   async recoverWallet(mnemonic: string, password: string): Promise<Wallet> {
     await RNZumoKit.recoverWallet(mnemonic, password);
     return new Wallet();
   }
 
+  /**
+   * Unlock user wallet with user's password.
+   * @param  password       user provided password
+   */
   async unlockWallet(password: string) {
     await RNZumoKit.unlockWallet(password);
     return new Wallet();
   }
 
+  /**
+   * Reveal mnemonic seed phrase used to seed user wallet.
+   * @param  password       user provided password
+   */
   async revealMnemonic(password: string): Promise<string> {
     return RNZumoKit.revealMnemonic(password);
   }
 
+  /**
+   * Check if mnemonic seed phrase corresponds to user's wallet.
+   * This is useful for validating seed phrase before trying to recover wallet.
+   * @param  mnemonic       mnemonic seed phrase
+   */
   async isRecoveryMnemonic(mnemonic: string): Promise<boolean> {
     return RNZumoKit.isRecoveryMnemonic(mnemonic);
   }
 
+  /**
+   * Get account in specific currency, on specific network, with specific type.
+   * @param  currencyCode   currency code, e.g. 'BTC', 'ETH' or 'GBP'
+   * @param  network        network type, e.g. 'MAINNET', 'TESTNET' or 'RINKEBY'
+   * @param  type           account type, e.g. 'STANDARD', 'COMPATIBILITY' or 'SEGWIT'
+   */
   getAccount(currencyCode: CurrencyCode, network: Network, type: AccountType) {
     return this.accounts.find(
       (account) =>
@@ -85,6 +137,10 @@ export class User implements IUser {
     );
   }
 
+  /**
+   * Check if user is a fiat customer on 'MAINNET' or 'TESTNET' network.
+   * @param  network 'MAINNET' or 'TESTNET'
+   */
   async isFiatCustomer(network: Network): Promise<boolean> {
     return RNZumoKit.isFiatCustomer(network);
   }
