+++ src/ZumoKit.ts
@@ -1,11 +1,13 @@
 import { NativeModules, NativeEventEmitter } from 'react-native';
-import { ExchangeRate } from 'zumokit/src/models/ExchangeRate';
-import { ExchangeRates } from 'zumokit/src/models/ExchangeRates';
-import { ExchangeSetting } from 'zumokit/src/models/ExchangeSetting';
-import { ExchangeSettings } from 'zumokit/src/models/ExchangeSettings';
-import { TransactionFeeRate } from 'zumokit/src/models/TransactionFeeRate';
-import { TransactionFeeRates } from 'zumokit/src/models/TransactionFeeRates';
-import { HistoricalExchangeRates } from 'zumokit/src/models/HistoricalExchangeRates';
+import {
+  ExchangeRate,
+  ExchangeRates,
+  ExchangeSetting,
+  ExchangeSettings,
+  TransactionFeeRate,
+  TransactionFeeRates,
+  HistoricalExchangeRates,
+} from 'zumokit/src/models';
 import {
   CurrencyCode,
   TokenSet,
@@ -14,37 +16,54 @@ import {
   ExchangeSettingJSON,
   TransactionFeeRateJSON,
   Dictionary,
-} from 'zumokit/src/types';
-import { ZumoKit as IZumoKit } from '../interfaces';
+} from 'zumokit/src/interfaces';
 import { Utils } from './Utils';
 import { User } from './User';
-import { tryCatchProxy } from '../utility/errorProxy';
+import { tryCatchProxy } from './utility/errorProxy';
 
 const {
   /** @internal */
   RNZumoKit,
 } = NativeModules;
 
+/**
+ * ZumoKit instance.
+ * <p>
+ * See <a href="https://developers.zumo.money/docs/guides/getting-started">Getting Started</a> guide for usage details.
+ * */
 @tryCatchProxy
-class ZumoKit implements IZumoKit {
+class ZumoKit {
   // The emitter that bubbles events from the native side
   private emitter = new NativeEventEmitter(RNZumoKit);
 
   // Listeners for exchange rates, exchange settings and transaction fee rates changes
   private changeListeners: Array<() => void> = [];
 
+  /** ZumoKit SDK semantic version tag if exists, commit hash otherwise. */
   version: string = RNZumoKit.version;
 
+  /** Currently signed-in user or null. */
   currentUser: User | null = null;
 
+  /** Crypto utilities. */
   utils: Utils = new Utils();
 
+  /** Mapping between currency pairs and available exchange rates. */
   exchangeRates: ExchangeRates = {};
 
+  /** Mapping between currency pairs and available exchange settings. */
   exchangeSettings: ExchangeSettings = {};
 
+  /** Mapping between cryptocurrencies and available transaction fee rates. */
   transactionFeeRates: TransactionFeeRates = {};
 
+  /**
+   * Initializes ZumoKit SDK. Should only be called once.
+   *
+   * @param apiKey        ZumoKit Api-Key
+   * @param apiUrl        ZumoKit API url
+   * @param txServiceUrl  ZumoKit Transaction Service url
+   */
   init(apiKey: string, apiUrl: string, txServiceUrl: string) {
     RNZumoKit.init(apiKey, apiUrl, txServiceUrl);
 
@@ -54,6 +73,12 @@ class ZumoKit implements IZumoKit {
     });
   }
 
+  /**
+   * Signs in user corresponding to user token set. Sets current user to the newly signed in user.
+   * Refer to <a href="https://developers.zumo.money/docs/setup/server#get-zumokit-user-token">Server</a> guide for details on how to get user token set.
+   *
+   * @param tokenSet   user token set
+   */
   async signIn(userTokenSet: TokenSet) {
     const json = await RNZumoKit.signIn(JSON.stringify(userTokenSet));
     this.currentUser = new User(json);
@@ -61,11 +86,20 @@ class ZumoKit implements IZumoKit {
     return this.currentUser;
   }
 
+  /** Signs out current user. */
   async signOut() {
     await RNZumoKit.signOut();
     this.currentUser = null;
   }
 
+  /**
+   * Get exchange rate for selected currency pair.
+   *
+   * @param fromCurrency   currency code
+   * @param toCurrency     currency code
+   *
+   * @return exchange rate or null
+   */
   getExchangeRate(fromCurrency: CurrencyCode, toCurrency: CurrencyCode): ExchangeRate | null {
     return Object.keys(this.exchangeRates).includes(fromCurrency)
       ? ((this.exchangeRates[fromCurrency] as Dictionary<CurrencyCode, ExchangeRate>)[
@@ -74,6 +108,14 @@ class ZumoKit implements IZumoKit {
       : null;
   }
 
+  /**
+   * Get exchange setting for selected currency pair.
+   *
+   * @param fromCurrency   currency code
+   * @param toCurrency     currency code
+   *
+   * @return exchange setting or null
+   */
   getExchangeSetting(fromCurrency: CurrencyCode, toCurrency: CurrencyCode): ExchangeSetting | null {
     return Object.keys(this.exchangeSettings).includes(fromCurrency)
       ? ((this.exchangeSettings[fromCurrency] as Dictionary<CurrencyCode, ExchangeSetting>)[
@@ -82,21 +124,43 @@ class ZumoKit implements IZumoKit {
       : null;
   }
 
+  /**
+   * Get transaction fee rate for selected crypto currency.
+   *
+   * @param currency   currency code
+   *
+   * @return transaction fee rate or null
+   */
   getTransactionFeeRate(currency: CurrencyCode): TransactionFeeRate | null {
     return Object.keys(this.transactionFeeRates).includes(currency)
       ? (this.transactionFeeRates[currency] as TransactionFeeRate)
       : null;
   }
 
+  /**
+   * Fetch historical exchange rates for supported time intervals.
+   *
+   * @return historical exchange rates
+   */
   async fetchHistoricalExchangeRates(): Promise<HistoricalExchangeRates> {
     const historicalExchangeRatesJSON = (await RNZumoKit.fetchHistoricalExchangeRates()) as HistoricalExchangeRatesJSON;
     return HistoricalExchangeRates(historicalExchangeRatesJSON);
   }
 
+  /**
+   * Listen to changes in current user’s sign in state, exchange rates, exchange settings or transaction fee rates.
+   *
+   * @param listener interface to listen to changes
+   */
   addChangeListener(listener: () => void) {
     this.changeListeners.push(listener);
   }
 
+  /**
+   * Remove change listener.
+   *
+   * @param listener interface to listen to changes
+   */
   removeChangeListener(listener: () => void) {
     let index = this.changeListeners.indexOf(listener);
     while (index !== -1) {
@@ -124,5 +188,7 @@ class ZumoKit implements IZumoKit {
     this.transactionFeeRates = TransactionFeeRates(tranactionFeeRatesJSON);
   }
 }
+
+export { ZumoKit };
 
 export default new ZumoKit();
