+++ src/ZumoKitError.ts
@@ -3,7 +3,7 @@
  * Refer to <a href="https://developers.zumo.money/docs/guides/handling-errors">Handling Errors</a>
  * guide for details on handling errors.
  */
-export interface ZumoKitError extends Error {
+class ZumoKitError extends Error {
   /**
    * Error type, such as api_connection_error, api_error, wallet_error etc.
    */
@@ -19,4 +19,16 @@ export interface ZumoKitError extends Error {
    * Error message.
    */
   message: string;
+
+  /** @internal */
+  constructor(error: { code: string; message: string; userInfo?: { type: string } }) {
+    const { code, message, userInfo } = error;
+
+    super(message);
+    this.type = userInfo && userInfo.type ? userInfo.type : 'zumo_kit_error';
+    this.code = code;
+    this.message = message;
+  }
 }
+
+export { ZumoKitError };
