--- src/interfaces/index.ts
@@ -1,22 +1,0 @@
-export { Account } from 'zumokit';
-export { AccountCryptoProperties } from 'zumokit';
-export { AccountDataSnapshot } from 'zumokit';
-export { AccountFiatProperties } from 'zumokit';
-export { ComposedExchange } from 'zumokit';
-export { ComposedTransaction } from 'zumokit';
-export { Exchange } from 'zumokit';
-export { ExchangeRate } from 'zumokit';
-export { ExchangeRates } from 'zumokit';
-export { ExchangeSetting } from 'zumokit';
-export { ExchangeSettings } from 'zumokit';
-export { Transaction } from 'zumokit';
-export { TransactionCryptoProperties } from 'zumokit';
-export { TransactionFiatProperties } from 'zumokit';
-export { TransactionFeeRate } from 'zumokit';
-export { TransactionFeeRates } from 'zumokit';
-export { HistoricalExchangeRates } from 'zumokit';
-export { User } from './User';
-export { Utils } from './Utils';
-export { Wallet } from './Wallet';
-export { ZumoKit } from './ZumoKit';
-export { ZumoKitError } from './ZumoKitError';
