--- src/types/exported.ts
@@ -1,13 +1,0 @@
-export {
-  Dictionary,
-  CurrencyType,
-  CurrencyCode,
-  Network,
-  AccountType,
-  TransactionType,
-  TransactionStatus,
-  ExchangeStatus,
-  TimeInterval,
-  TokenSet,
-  FiatCustomerData,
-} from 'zumokit/src/types/exported';
