--- src/types/index.ts
@@ -1,32 +1,0 @@
-export {
-  Dictionary,
-  CurrencyType,
-  CurrencyCode,
-  Network,
-  AccountType,
-  TransactionType,
-  TransactionStatus,
-  ExchangeStatus,
-  TimeInterval,
-  TokenSet,
-  FiatCustomerData,
-} from './exported';
-
-export {
-  AccountCryptoPropertiesJSON,
-  AccountFiatPropertiesJSON,
-  AccountJSON,
-  ExchangeRateJSON,
-  ExchangeSettingJSON,
-  ComposedTransactionJSON,
-  ComposedExchangeJSON,
-  TransactionFeeRateJSON,
-  TransactionCryptoPropertiesJSON,
-  TransactionFiatPropertiesJSON,
-  TransactionJSON,
-  ExchangeJSON,
-  AccountDataSnapshotJSON,
-  HistoricalExchangeRatesJSON,
-  ZumoKitErrorJSON,
-  UserJSON,
-} from './internal';
