+++ src/utility/errorProxy.ts
@@ -1,4 +1,4 @@
-import { ZumoKitError } from '../models/ZumoKitError';
+import { ZumoKitError } from '../ZumoKitError';
 
 /** @internal */
 function handler(fun: any) {
